+++ src/graphics/examples/CMakeLists.txt
@@ -1,3 +1,3 @@
-find_package(SFML 2.4.2 COMPONENTS window REQUIRED)
+find_package(SFML 2 COMPONENTS window REQUIRED)
 add_subdirectory(SimpleRectangleDraw)
 add_subdirectory(SimpleMovingRectangle)
