+++ src/graphics/src/CMakeLists.txt
@@ -1,17 +1,15 @@
-find_package(SFML 2.4.2 COMPONENTS system graphics REQUIRED)
-
-set(CPP_DIR graphics/details)
+find_package(SFML 2 COMPONENTS system graphics REQUIRED)
 
 add_library(graphics
-  ${CPP_DIR}/RendererPoolFactory.cpp
-  ${CPP_DIR}/RendererPoolSfml.cpp
-  ${CPP_DIR}/RenderTarget.cpp
-  ${CPP_DIR}/RendererIdGenerator.cpp
-  ${CPP_DIR}/SfmlRectangleShape.cpp)
+  RendererPoolFactory.cpp
+  RendererPoolSfml.cpp
+  RenderTarget.cpp
+  RendererIdGenerator.cpp
+  SfmlRectangleShape.cpp)
 
 target_include_directories(graphics
-  PUBLIC ${CMAKE_CURRENT_LIST_DIR}
-  PRIVATE ${CMAKE_CURRENT_LIST_DIR}/graphics)
+  PUBLIC ${CMAKE_CURRENT_LIST_DIR}/../includes
+  PRIVATE ${CMAKE_CURRENT_LIST_DIR})
 
 target_include_directories(graphics SYSTEM
   PRIVATE ${SFML_INCLUDE_DIR})
@@ -20,12 +18,12 @@ target_link_libraries(graphics
   PRIVATE ${SFML_LIBRARIES})
 
 add_executable(graphics-ut
-  ${CPP_DIR}/RendererPoolSfml.ut.cpp
-  ${CPP_DIR}/RenderTarget.ut.cpp
-  ${CPP_DIR}/SfmlRectangleShape.ut.cpp)
+  RendererPoolSfml.ut.cpp
+  RenderTarget.ut.cpp
+  SfmlRectangleShape.ut.cpp)
 
 target_include_directories(graphics-ut SYSTEM BEFORE
-  PRIVATE ${CMAKE_CURRENT_LIST_DIR}/graphics
+  PRIVATE ${CMAKE_CURRENT_LIST_DIR}
           ${UT_INCLUDES}
           ${SFML_INCLUDE_DIR})
