+++ src/graphics/src/RendererIdGenerator.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "RendererId.hpp"
+#include "graphics/RendererId.hpp"
 
 namespace graphics
 {
