+++ src/graphics/src/RendererPoolFactory.cpp
@@ -1,4 +1,4 @@
-#include "RendererPoolFactory.hpp"
+#include "graphics/RendererPoolFactory.hpp"
 
 #include "RendererPoolSfml.hpp"
 #include "RenderTarget.hpp"
