+++ src/graphics/src/RendererPoolSfml.cpp
@@ -2,7 +2,7 @@
 
 #include <SFML/Graphics/Color.hpp>
 #include <SFML/Graphics/CircleShape.hpp>
-#include <boost/range/algorithm/for_each.hpp>
+#include <range/v3/algorithm/for_each.hpp>
 
 #include <boost/uuid/uuid_io.hpp>
 
@@ -23,23 +23,17 @@ RendererId RendererPoolSfml::take(const Size& size, const Position& position)
     auto id = renderer_id_generator->generate();
     auto shape = shapes.emplace(id, sf::Vector2f{size.width, size.height});
     shape.first->second.setPosition(position.x, position.y);
-    std::clog << "x: " << position.x << ", y: " << position.y << '\n';
     return id;
 }
 
 void RendererPoolSfml::give_back(const RendererId& id)
 {
     trash.emplace(id);
-    std::clog << "give_back \n";
 }
 
 void RendererPoolSfml::cleanup_unused()
 {
-    for (const auto& id : trash)
-    {
-        shapes.erase(id);
-        std::clog << "left: " << shapes.size() << '\n';
-    }
+    ranges::for_each(trash, [this](const auto& id) { shapes.erase(id); });
     trash.clear();
 }
 
@@ -48,7 +42,7 @@ void RendererPoolSfml::render_all()
     cleanup_unused();
     context_renderer->clear(sf::Color::Black);
 
-    boost::for_each(shapes, [&](const auto& shape) {
+    ranges::for_each(shapes, [&](const auto& shape) {
         context_renderer->draw(shape.second);
     });
 }
